Fix build on clang

+++ src/something_sound.cpp
@@ -57,9 +57,9 @@ void sample_mixer_audio_callback(void *userdata, Uint8 *stream, int len)
     Sample_Mixer *mixer = (Sample_Mixer *)userdata;
 
     int16_t *output = (int16_t *)stream;
-    size_t output_len = len / sizeof(*output);
+    size_t output_len = (size_t) len / sizeof(*output);
 
-    memset(stream, 0, len);
+    memset(stream, 0, (size_t) len);
     for (size_t i = 0; i < SAMPLE_MIXER_CAPACITY; ++i) {
         for (size_t j = 0; j < output_len; ++j) {
             int16_t x = 0;
